limit number of tracked packet number ranges that need to be acked
This removes a potential DoS vector and it ensures that the size of
outgoing ACK frames does not get out of hand, which could cause a
deadlock in case ACK frames cannot fit in generated packets anymore.

Older packet number ranges are simply removed from the start of the
list, to keep the total list length under the limit. This will cause the
peer to spuriously retransmit the data within those packets, though that
seems better than the alternative.

Fixes #489.